refactor: clean up index exports
