+++ src/decorators/index.ts
@@ -1,4 +1,5 @@
 export { AutoRegister } from "./autoRegister";
+export type { ClassDecorator, ClassFieldDecorator } from "./decorators";
 export { Inject } from "./inject";
 export { Injectable } from "./injectable";
 export { InjectAll } from "./injectAll";
