+++ src/index.ts
@@ -1,7 +1,7 @@
 export type { Container, ContainerOptions } from "./container";
 export { createContainer } from "./container";
+export type { ClassDecorator, ClassFieldDecorator } from "./decorators";
 export { AutoRegister, Inject, Injectable, InjectAll, Scoped } from "./decorators";
-export type { ClassDecorator, ClassFieldDecorator } from "./decorators/decorators";
 export { inject, injectBy } from "./inject";
 export { injectAll } from "./injectAll";
 export { Injector } from "./injector";
